added fib4 using memoize crate

+++ chapter1/fib2/src/main.rs
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-fn fib2(n: u32) -> u32 {
+fn fib2(n: usize) -> usize {
     if n<2 {
         // println!("fib({}) -> {}",n,n);
         n
